+++ src/common/js/constants.js
@@ -35,5 +35,6 @@ Object.keys( Editor.C.Mode ).every(( k ) => {
 // the cookie name for the switch state, defined here to be available anywhere
 COOKIE_RESPONSIBLE = 'pwix:Editor';
 COOKIE_SWITCH_STATE = 'switch_state';
+COOKIE_SWITCH_STATE_IDENTIFIER = COOKIE_RESPONSIBLE + '/' + COOKIE_SWITCH_STATE + '/';
 
 I18N = 'pwix:editor:i18n';